[Refactor] (test) Move all testing items of enum objects to new testing module 'test.unit_test.model.metadata_enum'.
