+++ test/unit_test/model/metadata.py
@@ -1,77 +1,11 @@
-from smoothcrawler_cluster.model.metadata import CrawlerStateRole, TaskResult, State, Task, Heartbeat
+from smoothcrawler_cluster.model.metadata import State, Task, Heartbeat
+from smoothcrawler_cluster.model import CrawlerStateRole, TaskResult
 
-from typing import List, Dict, Callable, TypeVar
-from enum import Enum
+from typing import List, Dict, Callable
 from abc import ABCMeta
 import traceback
 import pytest
 import random
-
-
-_EnumT = TypeVar("_EnumT", bound=Enum)
-
-
-class _EnumObjTest(metaclass=ABCMeta):
-
-    def _run_enum_value_test(self, under_test_enum: Enum, expected_value: str) -> None:
-        assert under_test_enum.value == expected_value, f"The value of enum member '{under_test_enum}' should be '{expected_value}'."
-
-
-
-class TestCrawlerStateRole(_EnumObjTest):
-    """Test for the enum object key-value mapping."""
-
-    def test_runner_value(self) -> None:
-        _under_test_enum = CrawlerStateRole.Runner
-        _expected_value = "runner"
-        self._run_enum_value_test(_under_test_enum, _expected_value)
-
-
-    def test_backup_runner_value(self) -> None:
-        _under_test_enum = CrawlerStateRole.Backup_Runner
-        _expected_value = "backup-runner"
-        self._run_enum_value_test(_under_test_enum, _expected_value)
-
-
-    def test_dead_runner_value(self) -> None:
-        _under_test_enum = CrawlerStateRole.Dead_Runner
-        _expected_value = "dead-runner"
-        self._run_enum_value_test(_under_test_enum, _expected_value)
-
-
-    def test_dead_backup_runner_value(self) -> None:
-        _under_test_enum = CrawlerStateRole.Dead_Backup_Runner
-        _expected_value = "dead-backup-runner"
-        self._run_enum_value_test(_under_test_enum, _expected_value)
-
-
-
-class TestTaskResult(_EnumObjTest):
-    """Test for the enum object key-value mapping."""
-
-    def test_processing_value(self) -> None:
-        _under_test_enum = TaskResult.Processing
-        _expected_value = "processing"
-        self._run_enum_value_test(_under_test_enum, _expected_value)
-
-
-    def test_done_value(self) -> None:
-        _under_test_enum = TaskResult.Done
-        _expected_value = "done"
-        self._run_enum_value_test(_under_test_enum, _expected_value)
-
-
-    def test_terminate_value(self) -> None:
-        _under_test_enum = TaskResult.Terminate
-        _expected_value = "terminate"
-        self._run_enum_value_test(_under_test_enum, _expected_value)
-
-
-    def test_error_value(self) -> None:
-        _under_test_enum = TaskResult.Error
-        _expected_value = "error"
-        self._run_enum_value_test(_under_test_enum, _expected_value)
-
 
 
 class _MetaDataTest(metaclass=ABCMeta):
@@ -134,14 +68,12 @@ class _MetaDataTest(metaclass=ABCMeta):
             assert False, f"It should work finely without any issue.\n The error is: {traceback.format_exc()}"
 
 
-
 class TestState(_MetaDataTest):
     """Test for all the attributes of **State**."""
 
     @pytest.fixture(scope="function")
     def state(self) -> State:
         return State()
-
 
     def test_set_role_correctly(self, state: State) -> None:
         """
@@ -173,7 +105,6 @@ class TestState(_MetaDataTest):
             assert True, "It works finely."
             assert state.role == "runner", "The value should be same as it set."
 
-
     def test_set_role_incorrectly(self, state: State) -> None:
         """
         Test for setting the property *role* with invalid string value.
@@ -190,7 +121,6 @@ class TestState(_MetaDataTest):
             assert state.role is None, "The value should be None because it got fail when it set the value."
         else:
             assert False, f"It should work finely without any issue.\n The error is: {traceback.format_exc()}"
-
 
     def test_total_crawler(self, state: State) -> None:
         """
@@ -214,7 +144,6 @@ class TestState(_MetaDataTest):
             invalid_2_value=5.5
         )
 
-
     def test_total_runner(self, state: State) -> None:
         """
         Test for the property *total_runner* of **State**.
@@ -236,7 +165,6 @@ class TestState(_MetaDataTest):
             invalid_1_value="5",
             invalid_2_value=5.5
         )
-
 
     def test_total_backup(self, state: State) -> None:
         """
@@ -260,7 +188,6 @@ class TestState(_MetaDataTest):
             invalid_2_value=5.5
         )
 
-
     def test_current_crawler(self, state: State) -> None:
         """
         Test for the property *current_crawler* of **State**.
@@ -282,7 +209,6 @@ class TestState(_MetaDataTest):
             invalid_1_value="5",
             invalid_2_value={"k1": "v1", "k2": "v2", "k3": "v3"}
         )
-
 
     def test_current_runner(self, state: State) -> None:
         """
@@ -306,7 +232,6 @@ class TestState(_MetaDataTest):
             invalid_2_value={"k1": "v1", "k2": "v2", "k3": "v3"}
         )
 
-
     def test_current_backup(self, state: State) -> None:
         """
         Test for the property *current_backup* of **State**.
@@ -328,7 +253,6 @@ class TestState(_MetaDataTest):
             invalid_1_value="5",
             invalid_2_value={"k1": "v1", "k2": "v2", "k3": "v3"}
         )
-
 
     def test_standby_id(self, state: State) -> None:
         """
@@ -352,7 +276,6 @@ class TestState(_MetaDataTest):
             invalid_2_value=["iron_man_1"]
         )
 
-
     def test_fail_crawler(self, state: State) -> None:
         """
         Test for the property *fail_crawler* of **State**.
@@ -375,7 +298,6 @@ class TestState(_MetaDataTest):
             invalid_2_value={"k1": "v1", "k2": "v2", "k3": "v3"}
         )
 
-
     def test_fail_runner(self, state: State) -> None:
         """
         Test for the property *fail_runner* of **State**.
@@ -397,7 +319,6 @@ class TestState(_MetaDataTest):
             invalid_1_value="5",
             invalid_2_value={"k1": "v1", "k2": "v2", "k3": "v3"}
         )
-
 
     def test_fail_backup(self, state: State) -> None:
         """
@@ -422,14 +343,12 @@ class TestState(_MetaDataTest):
         )
 
 
-
 class TestTask(_MetaDataTest):
     """Test for all the attributes of **Task**."""
 
     @pytest.fixture(scope="function")
     def task(self) -> Task:
         return Task()
-
 
     def test_task_content(self, task: Task) -> None:
 
@@ -446,7 +365,6 @@ class TestTask(_MetaDataTest):
             invalid_1_value="5",
             invalid_2_value=["spider_1"]
         )
-
 
     def test_task_result(self, task: Task) -> None:
 
@@ -465,14 +383,12 @@ class TestTask(_MetaDataTest):
         )
 
 
-
 class TestHeartbeat(_MetaDataTest):
     """Test for all the attributes of **Heartbeat**."""
 
     @pytest.fixture(scope="function")
     def heartbeat(self) -> Heartbeat:
         return Heartbeat()
-
 
     def test_datetime(self, heartbeat: Heartbeat) -> None:
